# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: First-party Cognirise website and mini CMS API.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /submissions:
    get:
      operationId: listSubmissions
      tags: [submissions]
      summary: List enquiries and newsletter subscriptions
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/Search"
        - name: kind
          in: query
          schema:
            $ref: "#/components/schemas/SubmissionKind"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/SubmissionStatus"
        - name: market
          in: query
          schema: { type: string }
        - name: from
          in: query
          schema: { type: string, format: date-time }
        - name: to
          in: query
          schema: { type: string, format: date-time }
      responses:
        "200":
          description: Paginated submissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionPage"
  /submissions/export:
    post:
      operationId: exportSubmissions
      tags: [submissions]
      summary: Export filtered submissions
      security:
        - cookieAuth: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmissionExportInput"
      responses:
        "202":
          description: Export prepared or queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionExport"
  /submissions/{submissionId}:
    patch:
      operationId: updateSubmission
      tags: [submissions]
      summary: Update submission workflow state
      security:
        - cookieAuth: []
          csrfToken: []
      parameters:
        - $ref: "#/components/parameters/SubmissionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmissionUpdate"
      responses:
        "200":
          description: Submission updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
...
  /public/content:
    get:
      operationId: listPublishedContent
      tags: [public]
      summary: List published content
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: kind
          in: query
          schema:
            $ref: "#/components/schemas/DocumentKind"
        - name: market
          in: query
          required: true
          schema: { type: string }
        - name: locale
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Paginated published content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublishedContentPage"
  /public/content/{market}/{locale}/{kind}/{slug}:
    get:
      operationId: getPublishedContent
      tags: [public]
      summary: Get published content by kind and slug
      parameters:
        - name: market
          in: path
          required: true
          schema: { type: string }
        - name: locale
          in: path
          required: true
          schema: { type: string }
        - name: kind
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/DocumentKind"
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Published content
          headers:
            ETag:
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublishedContent"
        "404":
          $ref: "#/components/responses/NotFound"
  /public/sitemap:
    get:
      operationId: getPublicSitemap
      tags: [public]
      summary: Get public sitemap entries
      parameters:
        - name: market
          in: query
          schema: { type: string }
        - name: locale
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Sitemap entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sitemap"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: cognirise_session
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
  parameters:
    Page:
      name: page
      in: query
      schema: { type: number, multipleOf: 1, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      schema: { type: number, multipleOf: 1, minimum: 1, maximum: 100, default: 25 }
    Search:
      name: search
      in: query
      schema: { type: string, maxLength: 200 }
...
    SubmissionId:
      name: submissionId
      in: path
      required: true
      schema: { type: string }
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
...
  schemas:
    ApiError:
      type: object
      required: [error]
      properties:
        error: { type: string }
        code: { type: string }
        details:
          type: object
          additionalProperties: true
...
    DocumentKind:
      type: string
      enum: [person, partner, platform, publication, case-study]
...
    SeoMetadata:
      type: object
      required: [title, description]
      properties:
        title: { type: string }
        description: { type: string }
        canonicalUrl: { type: ["string", "null"] }
        noIndex: { type: boolean }
        imageId: { type: ["string", "null"] }
...
    SubmissionKind:
      type: string
      enum: [enquiry, newsletter]
    SubmissionStatus:
      type: string
      enum: [new, open, contacted, resolved, spam, unsubscribed]
    Submission:
      type: object
      required: [id, kind, email, market, status, sourcePage, consent, createdAt, updatedAt]
      properties:
        id: { type: string }
        kind:
          $ref: "#/components/schemas/SubmissionKind"
        name: { type: ["string", "null"] }
        email: { type: string, pattern: '^[^@\s]+@[^@\s]+\.[^@\s]+$' }
        organization: { type: ["string", "null"] }
        role: { type: ["string", "null"] }
        market: { type: string }
        processArea: { type: ["string", "null"] }
        challenge: { type: ["string", "null"] }
        status:
          $ref: "#/components/schemas/SubmissionStatus"
        ownerId: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        sourcePage: { type: string }
        consent: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    SubmissionUpdate:
      type: object
      additionalProperties: false
      properties:
        status:
          $ref: "#/components/schemas/SubmissionStatus"
        ownerId: { type: ["string", "null"] }
        notes: { type: ["string", "null"], maxLength: 4000 }
    SubmissionExportInput:
      type: object
      additionalProperties: false
      required: [format]
      properties:
        format: { type: string, enum: [csv] }
        kind:
          $ref: "#/components/schemas/SubmissionKind"
        status:
          $ref: "#/components/schemas/SubmissionStatus"
        market: { type: string }
        from: { type: string, format: date-time }
        to: { type: string, format: date-time }
    SubmissionExport:
      type: object
      required: [id, status, expiresAt]
      properties:
        id: { type: string }
        status: { type: string, enum: [queued, ready] }
        downloadUrl: { type: ["string", "null"] }
        expiresAt: { type: string, format: date-time }
    SubmissionPage:
      allOf:
        - $ref: "#/components/schemas/PageMetadata"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/Submission"
...
    PublishedContent:
      type: object
      required: [id, kind, slug, title, content, market, locale, revision, publishedAt, updatedAt]
      properties:
        id: { type: string }
        kind:
          $ref: "#/components/schemas/DocumentKind"
        slug: { type: string }
        title: { type: string }
        summary: { type: ["string", "null"] }
        content:
          type: object
          additionalProperties: true
        seo:
          $ref: "#/components/schemas/SeoMetadata"
        media:
          type: array
          items:
            $ref: "#/components/schemas/PublicMedia"
        market: { type: string }
        locale: { type: string }
        revision: { type: number, multipleOf: 1, minimum: 1 }
        publishedAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    PublicMedia:
      type: object
      required: [id, url, mimeType]
      properties:
        id: { type: string }
        url: { type: string }
        mimeType: { type: string }
        width: { type: ["number", "null"], multipleOf: 1 }
        height: { type: ["number", "null"], multipleOf: 1 }
        altText: { type: ["string", "null"] }
        caption: { type: ["string", "null"] }
        credit: { type: ["string", "null"] }
    PublishedContentPage:
      allOf:
        - $ref: "#/components/schemas/PageMetadata"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/PublishedContent"
    Sitemap:
      type: object
      required: [items, generatedAt]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/SitemapEntry"
        generatedAt: { type: string, format: date-time }
    SitemapEntry:
      type: object
      required: [url, updatedAt]
      properties:
        url: { type: string }
        updatedAt: { type: string, format: date-time }
        changeFrequency:
          type: string
          enum: [always, hourly, daily, weekly, monthly, yearly, never]
        priority: { type: number, minimum: 0, maximum: 1 }
        alternates:
          type: array
          items:
            $ref: "#/components/schemas/SitemapAlternate"
    SitemapAlternate:
      type: object
      required: [locale, url]
      properties:
        locale: { type: string }
        url: { type: string }
...
    PageMetadata:
      type: object
      required: [page, pageSize, total, totalPages]
      properties:
        page: { type: number, multipleOf: 1, minimum: 1 }
        pageSize: { type: number, multipleOf: 1, minimum: 1 }
        total: { type: number, multipleOf: 1, minimum: 0 }
        totalPages: { type: number, multipleOf: 1, minimum: 0 }